feat(codegen): implement MemoryLocal promotion and shadow-stack prologue/epilogue (issue #10)
Address-taken local variables are now promoted to MemoryLocal (shadow-stack
slots) instead of ScalarLocal. Functions that have any MemoryLocals gain a
shadow-stack frame: the prologue saves the stack pointer via local.tee and
decrements it by the total frame size; the epilogue restores it before every
explicit return and at the function's fallthrough exit.

Also updates all codegen unit tests to use the new WasmInstr struct API
introduced by the WasmVM refactor (opcode + 15-type Imm variant replaces the
165-type std::variant). A new instr_check.hpp helper header provides
is()/asI32Const()/asI64Const()/asOneIdx() wrappers that restore the
optional-pointer access style used in the tests.

+++ src/codegen/SymbolTable.hpp
@@ -18,7 +18,7 @@ struct ScalarLocal {
 
 struct MemoryLocal {
     wvmcc::parser::TypeNodePtr type;
-    int frameOffset; // Offset from frame pointer
+    size_t frameOffset; // Offset from frame pointer
 };
 
 struct GlobalScalar {
